% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dataline.R, R/set_dataline.R
\name{add_dataline}
\alias{add_dataline}
\alias{set_dataline}
\title{Set and add datalines of an scplot}
\usage{
add_dataline(object, variable, line, point, type = "continuous", ...)

set_dataline(object, variable, line, point, type, ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{variable}{Name of the dataline variable to apply the style.}

\item{line}{A list with datapoint attributes: colour,
linetype, size (see \code{\link[=element_line]{element_line()}}).}

\item{point}{A list with datapoint attributes: colour,
shape, size (see \code{\link[=element_point]{element_point()}}).}

\item{type}{Either "continuous" or "descrete"}

\item{...}{As a shortcut, arguments passed hear are bundled as \code{line}
arguments (see \code{\link[=element_line]{element_line()}}).}
}
\description{
Set and add datalines of an scplot
}
\examples{
p1 <- scplot(exampleAB_add) \%>\% add_dataline("depression", color =
"darkblue")
}
\seealso{
\code{\link[=element_line]{element_line()}}, \code{\link[=element_point]{element_point()}}
}
