% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ridge.R
\name{add_ridge}
\alias{add_ridge}
\title{Add a ridge to an scplot}
\usage{
add_ridge(object, color = "grey98", variable = ".dvar")
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{variable}{Name of the dataline variable to apply the style.}
}
\description{
Add a ridge to an scplot
}
