% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{add_text}
\alias{add_text}
\title{Add test to an scplot}
\usage{
add_text(
  object,
  label,
  case = 1,
  x,
  y,
  color = "black",
  size = 1,
  angle = 0,
  hjust = 0.5,
  vjust = 0.5
)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{label}{A Character vector with text labels.}

\item{case}{Numerical vector with the case number or character string "all"
for all cases.}

\item{x}{x position}

\item{y}{y position}
}
\description{
Add test to an scplot
}
