% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_title_caption.R
\name{add_title}
\alias{add_title}
\alias{add_caption}
\title{Add title and caption to an scplot}
\usage{
add_title(object, label, ...)

add_caption(object, label, ...)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{label}{A Character vector with text labels.}

\item{...}{Further styling arguments: color, size, face, family, hjust, vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}
}
\description{
Add title and caption to an scplot
}
