% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit_scplot.R
\name{.inherit_scplot}
\alias{.inherit_scplot}
\title{Dummy function to inherit global descriptions of parameters}
\usage{
.inherit_scplot(...)
}
\arguments{
\item{...}{Further styling arguments: color, size, face, family, hjust, vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}

\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{label}{A Character vector with text labels.}

\item{text}{A list with text styling arguments (color, size, face, family,
hjust, vjust, lineheight, angle).}

\item{background}{A list with background styling arguments (fill, color, size,
linetype).}

\item{case}{Numerical vector with the case number or character string "all"
for all cases.}

\item{variable}{Name of the dataline variable to apply the style.}
}
\description{
Dummy function to inherit global descriptions of parameters
}
