% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-scplot.R
\name{.merge_element}
\alias{.merge_element}
\title{Merge two theme elements}
\usage{
.merge_element(new, old)
}
\arguments{
\item{new}{A list or theme element with new values.}

\item{old}{A theme element to be updated.}
}
\value{
A merged theme element.
}
\description{
Merge two theme elements
}
