% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_theme.R
\name{extract_theme}
\alias{extract_theme}
\title{Extract theme from an scplot object}
\usage{
extract_theme(object)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}
}
\value{
An scplot-theme object
}
\description{
Extract theme from an scplot object
}
