% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetChroms.R
\name{GetChroms}
\alias{GetChroms}
\title{Get a list of the chromosomes of a genome}
\usage{
GetChroms(g_url, dl_folder, gen_name = NULL)
}
\arguments{
\item{g_url}{Url to the ".gtf" file of the researched genome on the ncbi website.}

\item{dl_folder}{Path to the folder to download the files in.}

\item{gen_name}{Name to store the downloaded files, and title name of the plot.}
}
\value{
List of chromosomes
}
\description{
Get a list of the chromosomes of a genome
}
\examples{
GetChroms(g_url = "ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/002/277/835/GCF_002277835.1_ASM227783v1/GCF_002277835.1_ASM227783v1_genomic.gtf.gz",
           dl_folder = "C:/Users/Downloads",
           gen_name = Flavobacterium_ir1)

}
