% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSequenceNow.R
\name{GetSequenceNow}
\alias{GetSequenceNow}
\title{Function to get the sequence of a gene in fasta format}
\usage{
GetSequenceNow(gen_name = NULL, dl_folder, gen_id, fasta_name,
  protein = FALSE)
}
\arguments{
\item{gen_name}{Name of the genome/folder. Has to be the same as in MakeGenPlot}

\item{dl_folder}{Path to the the download folder. Has to be the same as in MakeGenPlot}

\item{gen_id}{Name of the wanted gene (The gene name of a gene in a plot of MakeGenPlot when
show_gen_id = TRUE)}

\item{fasta_name}{The first line in the fasta file "> gen_id, strand = ... , fasta_name"}

\item{protein}{Logical. To get the protein sequence instead of the nucleotide sequence: protein = TRUE}
}
\value{
A fasta file with the sequence of the given gene.
}
\description{
Function to get the sequence of a gene in fasta format
}
\examples{
GetSequenceNow(gen_name = "Flavobacterium_ir1",
               dl_folder = "C:/Users/Downloads",
               gen_id = "B4N84_RS04735",
               fasta_name = "Bladibladibla",
               protein = TRUE)

}
