% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGenPlot.R
\name{MakeGenPlot}
\alias{MakeGenPlot}
\title{Make Genomic Plots Function}
\usage{
MakeGenPlot(g_url = NULL, f_url = NULL, dl_folder = NULL,
  chr_id = NULL, gen_name = NULL, start = 1, end = 10000,
  show_gene_id = FALSE)
}
\arguments{
\item{g_url}{Url to the ".gtf" file of the researched genome on the ncbi website.
when the url is the same as the .fna file (exept for the ".fna"), only the fnaUrl is enough.}

\item{f_url}{Url to the ".fna" file of the researched genome on the ncbi website.}

\item{dl_folder}{Path to the folder to download the files in.}

\item{chr_id}{Chromosome ID name of the researched chromosome, as named in
the gff and gtf files. As character. Default is the first chromosome.}

\item{gen_name}{Name to store the downloaded files, and title name of the plot.
The name of the organism without spaces is recommended. As character.}

\item{start}{Base start of plot.}

\item{end}{Base end of plot.}

\item{show_gene_id}{Show the gene ID instead of the product in the plot. Logical.
The gene ID is needed to get the sequence of a gene in function GetSequenceNow.}

\item{gene_only}{Show only the Gene or also start and stop codons? Logical.}
}
\value{
A plot of the inserted genome files.
}
\description{
Make Genomic Plots Function
}
\examples{
MakeGenPlot(g_url = "ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/002/277/835/GCF_002277835.1_ASM227783v1/GCF_002277835.1_ASM227783v1_genomic.gtf.gz",
           dl_folder = "C:/Users/Downloads",
           chr_id = "NZ_NQOT01000381.1",
           gen_name = "Flavobacterium_ir1",
           start = 80000,
           end = 90000,
           show_gene_id = FALSE,
           f_url = NULL)

}
