% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplots.R
\name{plot.sc_hplm}
\alias{plot.sc_hplm}
\title{This function generates a forest plot for the random effects of a mixed hplm model}
\usage{
\method{plot}{sc_hplm}(x, effect = "intercept", mark = "mean", ci = 0.95, ...)
}
\arguments{
\item{x}{The return from the \code{hplm()} function.}

\item{effect}{The specific effect to be plotted (default is the intercept).}

\item{mark}{Set a reference line.}

\item{ci}{Value between 0 and 1 for calculating the confidence interval.}

\item{...}{Further arguments.}
}
\value{
A forest plot displaying Tau-U effects.
}
\description{
This function generates a forest plot for the random effects of a mixed hplm model
}
\examples{
# plot(hplm(Leidig2018), effect = "intercept")

}
