% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sc_nap.R
\name{plot.sc_nap}
\alias{plot.sc_nap}
\title{Plot NAP Effects}
\usage{
\method{plot}{sc_nap}(x, effect = "rescaled", ...)
}
\arguments{
\item{x}{The return from the \code{nap()} function.}

\item{effect}{The specific effect to be plotted (default is "rescaled").}

\item{...}{Further arguments.}
}
\value{
A forest plot displaying Tau-U effects.
}
\description{
This function generates a forest plot of NAP effects.
}
\examples{
plot(tau_u(Leidig2018), effect = "A vs. B - Trend A")

}
