% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sc_rand.R
\name{plot.sc_rand}
\alias{plot.sc_rand}
\title{Plot Randomization Effects}
\usage{
\method{plot}{sc_rand}(
  object,
  type = "hist",
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  text_observed = "observed",
  color = "lightgrey",
  ...
)
}
\arguments{
\item{...}{Further arguments.}

\item{x}{The return from the \code{tau_u()} function.}

\item{effect}{The specific effect to be plotted (default is "A vs. B - Trend A").}
}
\value{
A forest plot displaying Tau-U effects.
}
\description{
This function generates a forest plot of Tau-U effects.
}
\examples{
# plot(tau_u(Leidig2018), effect = "A vs. B - Trend A")

}
