% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplots.R
\name{plot.sc_tauu}
\alias{plot.sc_tauu}
\title{Plot Tau-U Effects}
\usage{
\method{plot}{sc_tauu}(x, effect = "A vs. B - Trend A", ...)
}
\arguments{
\item{x}{The return from the \code{tau_u()} function.}

\item{effect}{The specific effect to be plotted (default is "A vs. B - Trend A").}

\item{...}{Further arguments.}
}
\value{
A forest plot displaying Tau-U effects.
}
\description{
This function generates a forest plot of Tau-U effects.
}
\examples{
# plot(tau_u(Leidig2018), effect = "A vs. B - Trend A")

}
