% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scplot.R
\name{print.scplot}
\alias{print.scplot}
\title{print.scplot
ggplot2 graphic plots from scdf files}
\usage{
\method{print}{scplot}(x, ...)
}
\arguments{
\item{object}{An scplot object}
}
\value{
A ggplot2 object
}
\description{
print.scplot
ggplot2 graphic plots from scdf files
}
