% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-export.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{unit}
\alias{element_line}
\alias{element_text}
\alias{element_rect}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:element]{element_line}}, \code{\link[ggplot2:element]{element_rect}}, \code{\link[ggplot2:element]{element_text}}, \code{\link[ggplot2:reexports]{unit}}}
}}

