% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot.sc_rand.R
\name{scplot.sc_rand}
\alias{scplot.sc_rand}
\title{Plot Randomization Effects}
\usage{
\method{scplot}{sc_rand}(object, type = "hist", add_density_curve = TRUE, ...)
}
\arguments{
\item{object}{The return from the \code{tau_u()} function.}

\item{type}{Either \code{"hist"} or \code{"xy"}.}

\item{add_density_curve}{If TRUE, adds a density curve to the histogram.}

\item{...}{Further arguments.}
}
\value{
A plot displaying the results of the randomization test.
}
\description{
This function generates plots of randomization test results.
}
\examples{
\dontrun{
res <- scan::rand_test(scan::exampleAB$Anja, limit = 1)
scplot(res, type = "hist")

scplot(res, type = "xy")
}
}
