% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot.sc_tauu.R
\name{scplot.sc_tauu}
\alias{scplot.sc_tauu}
\title{Plot Tau-U Effects}
\usage{
\method{scplot}{sc_tauu}(object, effect = 1, ...)
}
\arguments{
\item{object}{The return from the \code{tau_u()} function.}

\item{effect}{The specific effect to be plotted (default is "A vs. B - Trend A").}

\item{...}{Further arguments.}
}
\value{
A forest plot displaying Tau-U effects.
}
\description{
This function generates a forest plot of Tau-U effects.
}
\examples{
res <- scan::tau_u(scan::Leidig2018)
scplot(res, effect = 3)

}
