% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot_theme.R
\name{scplot_theme}
\alias{scplot_theme}
\title{Create a new theme for scplot}
\usage{
scplot_theme(theme, ...)
}
\arguments{
\item{theme, ...}{A character string with a predefined graphical theme as the basis.}
}
\description{
Create a new theme for scplot
}
