% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_phasenames.R
\name{set_phasenames}
\alias{set_phasenames}
\title{Set phasenames of an scplot}
\usage{
set_phasenames(object, labels = NULL, position = NULL, ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{labels}{A Character vector with text labels.}

\item{position}{Character string either 'left' or 'center'.}

\item{...}{Further styling arguments: color, size, face, family, hjust,
vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}
}
\description{
Set phasenames of an scplot
}
