% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theme_element.R
\name{set_theme_element}
\alias{set_theme_element}
\title{Set a theme element}
\usage{
set_theme_element(object, ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{...}{various style parameter}
}
\description{
Set a theme element
}
\details{
Usually, you don't need this function. Possible theme elements are.
\verb{"text", "plot.background", "panel.background", "panel.spacing.y", "dataline", "datapoint", "statline", "axis.line.x", "axis.line.y", "axis.ticks.length", "axis.ticks", "axis.title.y", "axis.title.x", "axis.text.x", "axis.text.y", "plot.title", "plot.caption", "plot.margin", "casenames", "casenames.strip", "casenames.position", "phasenames", "phasenames.position.x", "separators", "separators.extent", "label.text", "label.background", "label.padding", "grid", "legend.position", "legend.background", "legend.text", "legend.title", "legend.margin"}
}
