% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_axis_label.R
\name{set_xlabel}
\alias{set_xlabel}
\alias{set_ylabel}
\title{Set label for axis}
\usage{
set_xlabel(object, label = NULL, ...)

set_ylabel(object, label = NULL, ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{label}{A Character vector with text labels.}

\item{...}{Further styling arguments: color, size, face, family, hjust,
vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}
}
\description{
Set label for axis
}
