% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrow.R
\name{add_arrow}
\alias{add_arrow}
\title{Add arrrows to an scplot}
\usage{
add_arrow(
  object,
  case = 1,
  x0,
  y0,
  x1,
  y1,
  color = "black",
  angle = 30,
  length = unit(5, "points"),
  ends = "last",
  type = "open"
)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{case}{Numerical vector with the case number or character string "all"
for all cases.}

\item{x0}{Origin x position of the line.}

\item{y0}{Origin y position of the line.}

\item{x1}{End x position of the line.}

\item{y1}{End y position of the line.}

\item{length}{Size of the aroow angels.}
}
\description{
Add arrrows to an scplot
}
