% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_add_dataline.R
\name{add_dataline}
\alias{add_dataline}
\alias{set_dataline}
\title{Set and add datalines of an scplot}
\usage{
add_dataline(object, variable, color, width, linetype, dots, shape, size)

set_dataline(object, variable, color, width, linetype, dots, shape, size)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{variable}{Name of the dataline variable to apply the style.}

\item{color}{A character string or a number defining the color of an element.}

\item{width}{A number defining the width of the line.}

\item{linetype}{A character string with the line type: "solid", "dashed",
"dotted"}

\item{dots}{Color if datapoints. If set NULL, no points are plotted.}

\item{shape}{Number. See pch graphical parameter on par help page.}

\item{size}{Text size relative to the base text size.}
}
\description{
Set and add datalines of an scplot
}
