% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_add_dataline.R
\name{add_dataline}
\alias{add_dataline}
\alias{set_dataline}
\title{Set and add datalines of an scplot}
\usage{
add_dataline(object, variable, line, point, type = "continuous")

set_dataline(object, variable, line, point, type)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{variable}{Name of the dataline variable to apply the style.}

\item{line}{A list (class element_line) with datapoint attributes: colour, linetype, size.}

\item{point}{A list (class element_point) with datapoint attributes: colour, shape, size.}

\item{type}{Either "continuous" or "descrete"}
}
\description{
Set and add datalines of an scplot
}
