% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend.R
\name{add_legend}
\alias{add_legend}
\title{Add a legend to an scplot}
\usage{
add_legend(
  object,
  labels = NULL,
  section_labels = c("Lines", "Phases"),
  case = 1,
  position = "right",
  datalines = TRUE,
  statlines = TRUE,
  title = NULL,
  text = NULL,
  background = NULL
)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{labels}{A Character vector with text labels.}

\item{case}{Numerical vector with the case number or character string "all"
for all cases.}

\item{position}{The position ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{datalines}{If TRUE, a legend for the datalines is generated.}

\item{statlines}{If TRUE, a legend for the statlines is generated.}

\item{title}{A list with text style parameters for the title.}

\item{text}{A list with text styling arguments (color, size, face, family,
hjust, vjust, lineheight, angle).}

\item{background}{A list with background styling arguments (fill, color, size,
linetype).}
}
\description{
Add a legend to an scplot
}
