% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_marks.R
\name{add_marks}
\alias{add_marks}
\title{Add marks to an scplot}
\usage{
add_marks(
  object,
  case = 1,
  positions,
  color = "red",
  size = 1,
  shape = 1,
  variable = ".dvar"
)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{case}{Numerical vector with the case number or character string "all"
for all cases.}

\item{positions}{Either a vector indicating the dots to be highlighted or a
character string with a logical expression (e.g. values < mean(values))}

\item{shape}{Number. See pch graphical parameter on par help page.}

\item{variable}{Name of the dataline variable to apply the style.}
}
\description{
Add marks to an scplot
}
