% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ridge.R
\name{add_ridge}
\alias{add_ridge}
\title{Add a ridge to an scplot}
\usage{
add_ridge(object, color = "grey98", variable = ".dvar")
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{color}{A character string or a number defining the color of an element.}

\item{variable}{Name of the dataline variable to apply the style.}
}
\value{
An object of class \code{scplot} (see\code{\link[=scplot]{scplot()}}) with changed element
\code{ridges}.
}
\description{
Add a ridge to an scplot
}
