% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_statline.R
\name{add_statline}
\alias{add_statline}
\title{Add a statline to an scplot}
\usage{
add_statline(
  object,
  stat = c("mean", "median", "min", "max", "quantile", "sd", "mad", "trend", "trendA",
    "movingMean", "movingMedian", "loreg", "lowess", "loess"),
  phase = NULL,
  color = NULL,
  linewidth = NULL,
  size = NULL,
  linetype = NULL,
  variable = NULL,
  ...
)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{stat}{A character string for defining a statistical line or curve to be
plotted. Possible values: \verb{"median", "mean", "min", "max", "quantile", "trend", "trendA", "trendA_bisplit", "trendA_trisplit", "movingMean", "movingMedian", "loreg"}.}

\item{phase}{Either a numeric or a character vector specifying the reference
phase (see details)}

\item{color}{A character string or a number defining the color of an element.}

\item{linewidth}{A number with the width of the line.}

\item{size}{Text size relative to the base text size.}

\item{linetype}{A character string with the line type: "solid", "dashed",
"dotted"}

\item{variable}{Name of the dataline variable to apply the style.}

\item{...}{additional parameters passed to the statistical function.}
}
\description{
Add a statline to an scplot
}
\details{
The \code{phase} argument defines the reference phase for some
statistical functions (\verb{"median", "mean", "min", "max", "quantile"}). The
default is \code{NULL} which calculates and plots statistics for each phase
separately. The arguments takes a numeric vector (phase number(s)) or a
character vector (phase name(s)). When more than one phase is defines,
statistics are based on the combined values of these phases. Some of the
functions defined in \code{stats} have additional arguments. The \code{\link[=mean]{mean()}}
function has a trim argument (e.g. \code{trim = 0.1}). \code{\link[=quantile]{quantile()}} has a
proportion argument (e.g. \code{prob = 0.75} for calculating the 75\% quantile).
\code{trendA} has the argument \code{method} which takes \code{ols} or \code{theil-sen}.
\code{movingMean} and \code{movingMedian} have a lag argument (e.g. \code{lag = 2}). The
local-regression curve function \code{lowess} (or \code{loreg}) has a proportion
argument (e.g. \code{f = 0.5}) and the local-regression curve function \code{\link[=loess]{loess()}}
has a span argument (e.g. \code{span = 0.75}).
}
