% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_set_theme.R
\name{add_theme}
\alias{add_theme}
\alias{set_theme_element}
\title{Add and set the theme of an scplot}
\usage{
add_theme(object, theme, ...)

set_theme_element(object, ...)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{theme}{A character string with a predefined graphical theme.
Possible values: default, yaxis, tiny, small, big, chart, ridge,
annotate, grid, grid2, dark, nodots, sienna, phase_color, phase_shade}

\item{...}{various style parameter}
}
\description{
Add and set the theme of an scplot
}
