% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_theme.R
\name{add_theme}
\alias{add_theme}
\title{Add a theme of to an scplot}
\usage{
add_theme(object, theme, ...)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{theme}{A character string with a predefined graphical theme.}

\item{...}{Further styling arguments: color, size, face, family, hjust, vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}
}
\description{
Add a theme of to an scplot
}
