% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_caption.R
\name{add_title}
\alias{add_title}
\alias{add_caption}
\title{Add title and caption to an scplot}
\usage{
add_title(object, label, ...)

add_caption(object, label, ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{label}{A Character vector with text labels.}

\item{...}{List with text parameters (\verb{"family", "face", "colour", "size", "hjust", "vjust", "angle", "lineheight", "margin"}). See \code{\link[=element_text]{element_text()}}.}
}
\value{
An object of class \code{scplot} (see\code{\link[=scplot]{scplot()}}) with changed \code{title}
and \code{caption} elements.
}
\description{
Add title and caption to an scplot
}
