% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ggplot.R
\name{as_ggplot}
\alias{as_ggplot}
\title{Creates a ggplot2 object from an \code{\link[=scplot]{scplot()}} object}
\usage{
as_ggplot(scplot)
}
\arguments{
\item{scplot}{An \code{scplot} object}
}
\value{
A \code{ggplot2} plot object.
}
\description{
\code{as_ggplot()} takes an \code{scplot} object as input and returns a \code{ggplot2}
object.
}
\details{
\code{as_ggplot()} is used directly when you want to return a ggplot2 object for
further use with external ggplot functions.
}
