% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ggplot.R
\name{as_ggplot}
\alias{as_ggplot}
\title{Creates a ggplot2 object from an scplot object}
\usage{
as_ggplot(scplot)
}
\arguments{
\item{scplot}{An scplot object}
}
\value{
A ggplot2 object
}
\description{
Creates a ggplot2 object from an scplot object
}
\details{
`as_ggplot()` is used when you like to return a ggplot2 object for
further use with external ggplot functions.
}
