% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-scplot.R
\name{.check_theme}
\alias{.check_theme}
\title{Check theme object}
\usage{
.check_theme(theme)
}
\arguments{
\item{theme}{A theme object.}
}
\value{
The checked theme object.
}
\description{
Validates specific properties of a theme object.
}
\keyword{internal}
