% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit_scplot.R
\name{.inherit_scplot}
\alias{.inherit_scplot}
\title{Dummy function to inherit global descriptions of parameters}
\usage{
.inherit_scplot(
  object,
  label,
  labels,
  color,
  size,
  linewidth,
  hjust,
  vjust,
  angle,
  width,
  face,
  linetype,
  shape,
  text,
  background,
  case,
  variable,
  point,
  line,
  ...
)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{label}{A Character vector with text labels.}

\item{labels}{A Character vector with text labels.}

\item{color}{A character string or a number defining the color of an element.}

\item{size}{Text size relative to the base text size.}

\item{linewidth}{A number with the width of the line.}

\item{hjust}{Horizontal justification (in [0,1])}

\item{vjust}{Vertical justification (in [0,1])}

\item{angle}{Angle (in [0,360])}

\item{width}{Line width.}

\item{face}{Font face ("plain", "italic", "bold", "bold.italic")}

\item{linetype}{A character string with the line type: "solid", "dashed",
"dotted"}

\item{shape}{Number. See pch graphical parameter on par help page.}

\item{text}{A list with text styling arguments (color, size, face, family,
hjust, vjust, lineheight, angle).}

\item{background}{A list with background styling arguments (fill, color, size,
linetype).}

\item{case}{Numerical vector with the case number or character string "all"
for all cases.}

\item{variable}{Name of the dataline variable to apply the style.}

\item{point}{A list (class element_point) with datapoint attributes: colour, shape, size.}

\item{line}{A list (class element_line) with datapoint attributes: colour, linetype, size.}

\item{...}{Further styling arguments: color, size, face, family, hjust, vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}
}
\description{
Dummy function to inherit global descriptions of parameters
}
