% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-scplot.R
\name{.merge_theme}
\alias{.merge_theme}
\title{Merge two theme lists}
\usage{
.merge_theme(new, old)
}
\arguments{
\item{new}{A theme list with new values.}

\item{old}{A theme list to be updated.}
}
\value{
A merged theme list.
}
\description{
Merges two theme lists, updating the old theme with new values from the new theme.
}
\keyword{internal}
