% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-scplot.R
\name{.size}
\alias{.size}
\title{Convert relative size to absolute size}
\usage{
.size(x, base)
}
\arguments{
\item{x}{A size value. If relative size, use \code{rel()} function.}

\item{base}{Base size to convert relative size to absolute size.}
}
\value{
Absolute size value.
}
\description{
Convert relative size to absolute size
}
\keyword{internal}
