% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-scplot.R
\name{.upperfirst}
\alias{.upperfirst}
\title{Capitalize first letter of a string}
\usage{
.upperfirst(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with first letter capitalized.
}
\description{
Capitalize first letter of a string
}
\keyword{internal}
