% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_point.R
\name{element_point}
\alias{element_point}
\title{Point element}
\usage{
element_point(colour = NULL, size = NULL, shape = NULL, color = NULL)
}
\arguments{
\item{size}{Relative size.}

\item{shape}{Point shape.}

\item{color, colour}{Point colour..}
}
\description{
In conjunction with ggplot an object to represent point attributes.
}
