% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_theme.R
\name{extract_theme}
\alias{extract_theme}
\title{Extract theme from an scplot object}
\usage{
extract_theme(object)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}
}
\value{
An scplot-theme object
}
\description{
Extract theme from an scplot object
}
\examples{
my_theme <- scplot() \%>\%
  set_panel(color = "red") \%>\%
  set_base_text(size = 12, color = "blue") \%>\%
  set_dataline(color = "darkred", linewidth = 2) \%>\%
  extract_theme()
p1 <- scplot(exampleABC) \%>\% add_theme(my_theme)
}
