% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_theme.R
\name{new_theme}
\alias{new_theme}
\alias{extract_theme}
\title{Create a new scplot theme}
\usage{
new_theme()

extract_theme(object)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}
}
\value{
An scplot-theme object

An object of class \code{scplot-theme} which can be used with the
\code{\link[=set_theme]{set_theme()}} function.
}
\description{
Create a new scplot theme
}
\examples{
my_theme <- new_theme() \%>\%
  set_panel(color = "red") \%>\%
  set_base_text(size = 12, color = "blue") \%>\%
  set_dataline(color = "darkred", linewidth = 2)
p1 <- scplot(exampleABC) \%>\% set_theme(my_theme)
}
