% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sc_rand.R
\name{plot.sc_rand}
\alias{plot.sc_rand}
\title{Plot Randomization Effects}
\usage{
\method{plot}{sc_rand}(
  object,
  type = "hist",
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  text_observed = "observed",
  color = "lightgrey",
  add_density_curve = TRUE,
  ...
)
}
\arguments{
\item{type}{Either \code{"hist"} or \code{"xy"}.}

\item{...}{Further arguments.}

\item{x}{The return from the \code{tau_u()} function.}
}
\value{
A forest plot displaying Tau-U effects.
}
\description{
This function generates a forest plot of Tau-U effects.
}
\examples{
# plot(rand_test(exampleAB$Anja, limit = 1), type = "hist")
# plot(rand_test(exampleAB$Anja, limit = 1), type = "xy")

}
