% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sc_rand.R
\name{plot.sc_rand}
\alias{plot.sc_rand}
\title{Plot Randomization Effects}
\usage{
\method{plot}{sc_rand}(x, type = "hist", add_density_curve = TRUE, ...)
}
\arguments{
\item{x}{The return from the \code{tau_u()} function.}

\item{type}{Either \code{"hist"} or \code{"xy"}.}

\item{add_density_curve}{If TRUE, adds a density curve to the histogram.}

\item{...}{Further arguments.}
}
\value{
A forest plot displaying Tau-U effects.
}
\description{
This function generates a forest plot of Tau-U effects.
}
\examples{
res <- rand_test(exampleAB$Anja, limit = 1)
plot(res, type = "hist")

plot(res, type = "xy")

}
