% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scplot.R
\name{print.scplot}
\alias{print.scplot}
\title{print.scplot}
\usage{
\method{print}{scplot}(x, ...)
}
\arguments{
\item{x}{An scplot object}
}
\value{
Creates a single-case plot
}
\description{
print.scplot
}
