% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scplot.R
\name{print.scplot}
\alias{print.scplot}
\title{print.scplot}
\usage{
\method{print}{scplot}(x, ...)
}
\arguments{
\item{x}{An scplot object.}

\item{...}{Not implemented.}
}
\value{
Calls \code{\link[=as_ggplot]{as_ggplot()}} to create a ggplot2 argument and prints it.
}
\description{
print.scplot
}
\keyword{internal}
