% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot-package.R
\docType{package}
\name{scplot-package}
\alias{scplot-package}
\title{Single-Case Data Plots}
\description{
A collection of procedures for visualizing single-case data. It is an
add-on package for the \code{scan} package.
}
\author{
Juergen Wilbert [aut, cre]
}
\keyword{package}
