% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPlot-package.R
\docType{package}
\name{scPlot-package}
\alias{scPlot}
\alias{scPlot-package}
\title{scPlot: Plot}
\description{
Plot genomic ranges and get the sequences of genes.
}
\author{
\strong{Maintainer}: Boaz Kleijne \email{noobsugar@hotmail.com} (YOUR-ORCID-ID)

}
\keyword{internal}
