% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot.R, R/scplot_functions.R
\name{scplot}
\alias{scplot}
\alias{set_xlabel}
\alias{set_ylabel}
\alias{set_xaxis}
\alias{set_yaxis}
\alias{add_title}
\alias{add_caption}
\alias{set_casenames}
\alias{set_phasenames}
\alias{set_seperator}
\alias{add_grid}
\alias{add_labels}
\alias{add_dataline}
\alias{set_dataline}
\alias{add_statline}
\alias{set_background}
\alias{set_panel}
\alias{add_theme}
\alias{add_marks}
\alias{add_text}
\alias{add_arrow}
\alias{set_theme_element}
\alias{add_ridge}
\alias{add_legend}
\title{Plot single-case data}
\usage{
scplot(data)

set_xlabel(object, label = NULL, ...)

set_ylabel(object, label, ...)

set_xaxis(
  object,
  limits = NULL,
  increment = NULL,
  increment_from = NULL,
  positions = NULL,
  ...
)

set_yaxis(
  object,
  limits = NULL,
  increment = NULL,
  increment_from = NULL,
  positions = NULL,
  ...
)

add_title(object, label, ...)

add_caption(object, label, ...)

set_casenames(
  object,
  ...,
  x = NULL,
  y = NULL,
  type = NULL,
  text = list(),
  background = list()
)

set_phasenames(
  object,
  ...,
  color = NULL,
  size = NULL,
  x = NULL,
  y = NULL,
  text = list()
)

set_seperator(object, extent = NULL, label = NULL, ...)

add_grid(object, ...)

add_labels(
  object,
  nudge_y = NULL,
  nudge_x = NULL,
  round = NULL,
  text = list(),
  background = list(),
  variable = ".dvar"
)

add_dataline(object, variable, color, width, linetype, dots, shape, size)

set_dataline(object, variable, color, width, linetype, dots, shape, size)

add_statline(
  object,
  stat,
  phase = NULL,
  color = NULL,
  width = NULL,
  linetype = NULL,
  variable = NULL,
  ...
)

set_background(object, ...)

set_panel(object, alpha = NULL, ...)

add_theme(object, theme, ...)

add_marks(
  object,
  case = 1,
  positions,
  color = "red",
  size = 1,
  shape = 1,
  variable = ".dvar"
)

add_text(object, label, case = 1, x, y, color = "black", size = 2, angle = 0)

add_arrow(
  object,
  case = 1,
  x0,
  y0,
  x1,
  y1,
  color = "black",
  angle = 30,
  length = unit(5, "points"),
  ends = "last",
  type = "open"
)

set_theme_element(object, ...)

add_ridge(object, color = "grey98", variable = ".dvar")

add_legend(
  object,
  labels = ".default",
  case = 1,
  position = "right",
  datalines = TRUE,
  statlines = TRUE,
  title = "Lines"
)
}
\arguments{
\item{label}{Character string.}

\item{...}{various style parameter}

\item{limits}{Lower and upper limits of the axis (e.g., \code{limits = c(0,
20)} sets the axis to a scale from 0 to 20). With multiple single-cases
you can use \code{limits = c(0, NA)} to scale the axis from 0 to the maximum
of each case. \code{limits} is not set by default, which makes \code{scan} set
a proper scale based on the given data.}

\item{increment}{An integer. Increment of the x-axis. 1 :each mt value will be printed, 2 : every other value, 3 : every third values etc.}

\item{positions}{Either a vector indicating the dots to be highlighted or a
character string with a logical expression (e.g. values < mean(values))}

\item{x}{x position}

\item{y}{y position}

\item{color}{A character string or a number defining the color of an element.}

\item{extent}{A number between 0 and 1 given the proportion of the plot that is covert by the line or character string "full" or "scale".}

\item{round}{Number of digits of the labels.}

\item{width}{A number deifning the width of the line.}

\item{linetype}{A character string with the line type: "solid", "dashed", "dotted"}

\item{shape}{Number. See pch graphical parameter on par help page.}

\item{stat}{A character string for defining a statistical line or curve to
be plotted. Possible values: "median", "mean", "min", "max", "quantile",
"trend", "trendA",
"trendA_bisplit", "trendA_trisplit",
"movingMean", "movingMedian", "loreg".}

\item{phase}{Either a numeric or a character vector specifying the reference phase (see details)}

\item{theme}{A character string with a predefined graphical theme.
Possible values: default, yaxis, tiny, small, big, chart, ridge,
annotate, grid, grid2, dark, nodots, sienna, phase_color, phase_shade}

\item{case}{Numerical vector with the csae number or character string "all" for all cases.}

\item{x0}{Origin x position of the line.}

\item{y0}{Origin y position of the line.}

\item{x1}{End x position of the line.}

\item{y1}{End y position of the line.}

\item{length}{Size of the aroow angels.}

\item{orientation}{of the label: 0 = vertical; 1 = horizontal}

\item{align}{Character string. One of "left", "right", "center"}

\item{wrap}{Number that defines the maximum characters per line before a break.
If set to FALSE, no automatic linebreak is set.}

\item{parse}{If TRUE, the label is interpreted as an expression. Default = FALSE.}
}
\value{
An scplot object that creates a ggplot2 plot when printed.
}
\description{
This function provides a plot of a single-case or multiple
single-cases.
}
\details{
The 'phase' argument defines the reference phase for some statistiscal functions ("median", "mean", "min", "max", "quantile"). The default is NULL which calculates and plots statistics for each phase seperately. The arguments takes a numeric vector (phase number(s)) or a character vector (phase name(s)). When more than one phase is defines, statistics are based on the combined values of these phases.
Some of the functions defined in 'stats' have additional arguments.
The 'mean' function has a trim argument (e.g. 'trim = 0.1').
'quantile has a proportion argument (e.g. prob = 0.75 for calculating
the 75% quantile).
'movingMean' and 'movingMedian' have a lag argument (e.g. lag = 2).
The local-regression curve function 'lowess' (or 'loreg') has a proportion
argument (e.g. f = 0.5) and the
local-regression curve function 'loess' has a span argument
(e.g. span = 0.75).
}
\author{
Juergen Wilbert
}
