% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reworked_scplot_functions.R, R/scplot.R,
%   R/scplot_functions.R
\name{set_xlabel}
\alias{set_xlabel}
\alias{set_ylabel}
\alias{set_xaxis}
\alias{set_yaxis}
\alias{add_title}
\alias{add_caption}
\alias{set_casenames}
\alias{set_phasenames}
\alias{set_seperator}
\alias{add_grid}
\alias{add_labels}
\alias{add_dataline}
\alias{set_dataline}
\alias{add_statline}
\alias{set_background}
\alias{set_panel}
\alias{scplot}
\alias{add_theme}
\alias{set_theme_element}
\alias{add_marks}
\alias{add_text}
\alias{add_arrow}
\alias{add_ridge}
\alias{add_legend}
\alias{add_margins}
\title{Plot single-case data}
\usage{
set_xlabel(
  object,
  label = NULL,
  color = NULL,
  size = NULL,
  family = NULL,
  face = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  margin = NULL
)

set_ylabel(
  object,
  label,
  orientation,
  color = NULL,
  size = NULL,
  family = NULL,
  face = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  margin = NULL
)

set_xaxis(
  object,
  limits,
  increment,
  increment_from,
  color,
  size,
  line,
  positions
)

set_yaxis(object, limits, color, size, increment, increment_from, positions)

add_title(
  object,
  label,
  color = NULL,
  size = NULL,
  family = NULL,
  face = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  margin = NULL,
  parse = NULL
)

add_caption(
  object,
  label,
  color = NULL,
  size = NULL,
  face = NULL,
  hjust = NULL,
  vjust = NULL,
  wrap = NULL,
  margin = NULL,
  parse = NULL
)

set_casenames(
  object,
  ...,
  x = NULL,
  y = NULL,
  color = NULL,
  size = NULL,
  hjust = NULL,
  vjust = NULL,
  width = NULL,
  angle = NULL,
  fill = NULL,
  linetype = NULL,
  frame = NULL,
  type = NULL
)

set_phasenames(
  object,
  ...,
  color = NULL,
  size = NULL,
  x = NULL,
  y = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  box = NULL,
  frame = NULL,
  face = NULL,
  family = NULL
)

set_seperator(
  object,
  color = NULL,
  width = NULL,
  linetype = NULL,
  extent = NULL,
  label = NULL,
  size = NULL
)

add_grid(
  object,
  color = NULL,
  size = NULL,
  linetype = NULL,
  lineend = NULL,
  arrow = NULL
)

add_labels(
  object,
  color = NULL,
  size = NULL,
  nudge_y,
  nudge_x,
  vjust = NULL,
  hjust = NULL,
  angle = NULL,
  round,
  box = NULL,
  frame = NULL
)

add_dataline(object, variable, color, width, linetype, dots, shape, size)

set_dataline(object, variable, color, width, linetype, dots, shape, size)

add_statline(
  object,
  stat,
  color = NULL,
  width = NULL,
  linetype = NULL,
  variable = NULL
)

set_background(object, color = NULL, fill = NULL, linetype = NULL, size = NULL)

set_panel(object, color = NULL, fill = NULL, linetype = NULL, size = NULL)

scplot(data)

add_theme(object, theme, ...)

set_theme_element(object, ...)

add_marks(
  object,
  case = 1,
  positions,
  color = "red",
  size = 1,
  shape = 1,
  variable = ".dvar"
)

add_text(object, case = 1, x, y, label, color = NULL, size = NULL, angle = 0)

add_arrow(object, case = 1, x0, y0, x1, y1, length = 0.1, color = NULL, ...)

add_ridge(object, color = "grey98")

add_legend(
  object,
  labels = ".default",
  case = 1,
  x,
  y,
  datalines = TRUE,
  statlines = TRUE,
  title = NULL
)

add_margins(object, outer, inner)
}
\arguments{
\item{label}{Character string.}

\item{color}{A character string or a number defining the color of an element.}

\item{size}{A number deifning the size of an element.}

\item{orientation}{of the label: 0 = vertical; 1 = horizontal}

\item{limits}{Lower and upper limits of the axis (e.g., \code{limits = c(0,
20)} sets the axis to a scale from 0 to 20). With multiple single-cases
you can use \code{limits = c(0, NA)} to scale the axis from 0 to the maximum
of each case. \code{limits} is not set by default, which makes \code{scan} set
a proper scale based on the given data.}

\item{increment}{An integer. Increment of the x-axis. 1 :each mt value will be printed, 2 : every other value, 3 : every third values etc.}

\item{positions}{Either a vector indicating the dot to be highlighted or a
character string with a logical expression (e.g. values < mean(values))}

\item{parse}{If TRUE, the label is interpreted as an expression. Default = FALSE.}

\item{wrap}{Number that defines the maximum characters per line before a break.
If set to FALSE, no automatic linebreak is set.}

\item{...}{various style parameter}

\item{x}{x position}

\item{y}{y position}

\item{width}{A number defining the line width}

\item{type}{A character string with the line type: "solid", "dashed", "dotted"}

\item{extent}{A number between 0 and 1 given the proportion of the plot that is covert by the line or character string "full" or "scale".}

\item{round}{Number of digits of the labels.}

\item{shape}{Number. See pch graphical parameter on par help page.}

\item{stat}{A character string defining a line or curve to be
plotted. Possible values: "median", "mean", "trend", "trendA",
"trendA_bisplit", "trendA_trisplit", "maxA", "minA", "meanA", "medianA",
"plm", "movingMean", "movingMedian", "loreg"}

\item{theme}{A character string with a predefined graphical theme.
Possible values: default, yaxis, tiny, small, big, chart, ridge,
annotate, grid, grid2, dark, nodots, sienna, phase_color, phase_shade}

\item{case}{Numerical vector with the csae number or character string "all" for all cases.}

\item{x0}{Origin x position of the line.}

\item{y0}{Origin y position of the line.}

\item{x1}{End x position of the line.}

\item{y1}{End y position of the line.}

\item{length}{Size of the aroow angels.}

\item{outer}{Vector with four values for the extension of the inner margins
(negative numbers for smaller margins): c(bottom, left, top, right)}

\item{align}{Character string. One of "left", "right", "center"}
}
\value{
An scplot object that creates a ggplot2 plot when printed.
}
\description{
This function provides a plot of a single-case or multiple
single-cases.
}
\author{
Juergen Wilbert
}
