% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot.R
\name{scplot}
\alias{scplot}
\title{Plot single-case data}
\usage{
scplot(scdf)
}
\arguments{
\item{scdf}{A single-case data-frame object (scdf).}
}
\value{
An object of class \code{scplot} containing the single-case data (element \code{scdf}),
and information about the plot style (element \code{theme}).
}
\description{
This function provides a plot of a single-case or multiple single-cases.
}
\author{
Juergen Wilbert
}
