% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot_rand.R
\name{scplot_rand}
\alias{scplot_rand}
\title{Random start position plot
Plot of statistics for random phase B start positions}
\usage{
scplot_rand(
  scdf,
  statistic = "Mean B-A",
  colors = c("coral3", "aquamarine4", "#56B4E9", "black"),
  ...
)
}
\arguments{
\item{scdf}{A single-case data frame object.}

\item{statistic}{A string with a the name of a statistic.
Defaults to \verb{Mean B-A}. See rand_test() function for all options.}

\item{...}{further arguments passted to the scan \code{rand_test()} function.}
}
\description{
Random start position plot
Plot of statistics for random phase B start positions
}
\examples{
scplot_stats(scan::byHeart2011[1:5])

}
