% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_background.R
\name{set_background}
\alias{set_background}
\alias{set_panel}
\title{Set plot and panel background of an scplot}
\usage{
set_background(object, ...)

set_panel(object, ...)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{...}{Further styling arguments: color, size, face, family, hjust, vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}
}
\description{
Set plot and panel background of an scplot
}
