% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_panel.R
\name{set_background}
\alias{set_background}
\alias{set_panel}
\title{Set plot and panel background of an scplot}
\usage{
set_background(object, ...)

set_panel(object, ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{...}{Further styling arguments: color, size, face, family, hjust,
vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}
}
\description{
Set plot and panel background of an scplot
}
