% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_base_text.R
\name{set_base_text}
\alias{set_base_text}
\title{Set base text characteristics of an scplot}
\usage{
set_base_text(object, ...)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{...}{Further styling arguments: color, size, face, family, hjust, vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}
}
\description{
Set base text characteristics of an scplot
}
