% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_base_text.R
\name{set_base_text}
\alias{set_base_text}
\title{Set base text parameters of an scplot}
\usage{
set_base_text(object, ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{...}{List with text parameters (\verb{"family", "face", "colour", "size", "hjust", "vjust", "angle", "lineheight", "margin"}). See \code{\link[=element_text]{element_text()}}.}
}
\value{
An object of class \code{scplot} (see\code{\link[=scplot]{scplot()}}).
}
\description{
Set base text parameters of an scplot
}
