% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_casenames.R
\name{set_casenames}
\alias{set_casenames}
\title{Set casenames of an scplot}
\usage{
set_casenames(object, labels = NULL, position = NULL, ..., background = list())
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{...}{Further styling arguments: color, size, face, family, hjust, vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}

\item{background}{A list with background styling arguments (fill, color, size,
linetype).}
}
\description{
Set casenames of an scplot
}
