% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_casenames.R
\name{set_casenames}
\alias{set_casenames}
\title{Set casenames of an scplot}
\usage{
set_casenames(object, labels = NULL, position = NULL, background = list(), ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{labels}{A Character vector with text labels.}

\item{position}{Either \verb{"topleft", "bottomleft", "topright", "bottomright", "strip-right", "strip-top"}, or a numerical vector of length 2 with the x
and y position (e.g. \code{c(19, 20)}).}

\item{background}{A list with background styling arguments (fill, color,
size, linetype).}

\item{...}{List with text parameters (\verb{"family", "face", "colour", "size", "hjust", "vjust", "angle", "lineheight", "margin"}). See \code{\link[=element_text]{element_text()}}.}
}
\value{
An object of class \code{scplot} (see\code{\link[=scplot]{scplot()}}) with a changed \code{casenames}
element.
}
\description{
Set casenames of an scplot
}
