% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_phasenames.R
\name{set_phasenames}
\alias{set_phasenames}
\title{Set phasenames of an scplot}
\usage{
set_phasenames(object, labels = NULL, position = NULL, ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{labels}{A Character vector with text labels.}

\item{position}{Character string either 'left' or 'center'.}

\item{...}{List with text parameters (\verb{"family", "face", "colour", "size", "hjust", "vjust", "angle", "lineheight", "margin"}). See \code{\link[=element_text]{element_text()}}.}
}
\value{
An object of class \code{scplot} (see\code{\link[=scplot]{scplot()}}) with a changed \code{phasenames}
element.
}
\description{
Set phasenames of an scplot
}
