% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_separator.R
\name{set_separator}
\alias{set_separator}
\title{Set separator line in an scplot}
\usage{
set_separator(object, ...)
}
\arguments{
\item{object}{An scplot object (class `scplot`) returned from the `scplot()`
function.}

\item{...}{Further styling arguments: color, size, face, family, hjust, vjust, lineheight, angle, linetype, lineend, arrow, fill, margin.}
}
\description{
Set separator line in an scplot
}
