% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_separator.R
\name{set_separator}
\alias{set_separator}
\title{Set separator line in an scplot}
\usage{
set_separator(object, ...)
}
\arguments{
\item{object}{An scplot object (class \code{scplot}) returned from the \code{scplot()}
function.}

\item{...}{List with text parameters (\verb{"family", "face", "colour", "size", "hjust", "vjust", "angle", "lineheight", "margin"}). See \code{\link[=element_text]{element_text()}}.}
}
\description{
Set separator line in an scplot
}
