% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dataline.R
\name{split_dataline}
\alias{split_dataline}
\title{Split Treatment Dataline}
\usage{
split_dataline(
  object,
  tvar,
  dvar = NULL,
  levels = NULL,
  labels = NULL,
  remove_original = TRUE,
  color = NULL,
  ...
)
}
\arguments{
\item{object}{A scplot object containing the scdf data.}

\item{tvar}{The treatment variable.}

\item{dvar}{The dependent variable which provides the values for the new
datalines. If NULL, the default dependent variable of the scplot object is
used.}

\item{levels}{Optional, a vector of treatment levels.}

\item{labels}{Optional, labels for the treatment levels.}

\item{remove_original}{Logical, should the original dependent variable be
removed from the cases where the treatment variable matches the level?
(default is FALSE).}

\item{color}{Optional, a vector of colors for the new datalines. If provided,
the length of the color vector should match the number of treatment levels.
If not provided, default colors will be used for the datalines.}

\item{...}{Additional arguments.}
}
\value{
The modified scplot object with added datalines.
}
\description{
This function takes a scplot object and adds treatment datalines according to
the specified variables.
}
\details{
The function identifies unique treatment levels from the specified treatment
variable and creates new datalines for each level. The new datalines are
added to the scplot object, and the original dependent variable can
optionally be removed from cases where the treatment variable matches the
level.
}
